feat: add recomputeOrganizationBalances repair tool and fix tests
- Add recomputeOrganizationBalances analogous to the user version,
  replaying org usage/credit ledger to fix drift
- Fix BigInt/number mismatch in SQL computed balance by casting to float8
- Update all org test fixtures to set total_microdollars_acquired
  alongside microdollars_balance
- Add total_microdollars_acquired assertions to billing, promotions,
  admin, seats, and auto-top-up tests

+++ src/lib/organizations/organization-usage.ts
@@ -66,7 +66,7 @@ export async function getBalanceForOrganizationUser(
       microdollar_limit: organization_user_limits.microdollar_limit,
       microdollar_usage: organization_user_usage.microdollar_usage,
       organization_balance:
-        sql<number>`${organizations.total_microdollars_acquired} - ${organizations.microdollars_used}`.as(
+        sql<number>`(${organizations.total_microdollars_acquired} - ${organizations.microdollars_used})::float8`.as(
           'organization_balance'
         ),
       total_microdollars_acquired: organizations.total_microdollars_acquired,
@@ -214,7 +214,7 @@ export async function ingestOrganizationTokenUsage(usage: MicrodollarUsage): Pro
     const [orgData] = await tx
       .select({
         microdollars_balance:
-          sql<number>`${organizations.total_microdollars_acquired} - ${organizations.microdollars_used}`.as(
+          sql<number>`(${organizations.total_microdollars_acquired} - ${organizations.microdollars_used})::float8`.as(
             'microdollars_balance'
           ),
         settings: organizations.settings,
